Extract ThinkTag, we still need segments way.
